image/store/requirements.txt
pynamodb==1.5.1
python-magic==0.4.11
